requirements.txt
Flask==0.11.1
Flask-Bootstrap==3.3.6.0
Flask-WTF==0.12
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.3.1
WTForms==2.1
Werkzeug==0.11.10
click==6.6
dominate==2.2.1
flask-appconfig==0.11.1
itsdangerous==0.24
six==1.10.0
visitor==0.1.3
wsgiref==0.1.2
